[#19144, #19145, #19146]: DocDB, ASH: Instrument ycql, compactions and flushes
Summary:
Track wait states for cql requests, and also compaction and flush jobs.

Upgrade/Rollback safety:
Safe to upgrade/downgrade. Protos only affect the new functionality -- ASH, and
do not interfere with existing functionality. ASH will be unavailable if downgraded.
`TEST_yb_enable_ash` is used to guard this feature.

Test Plan:
ybd --cxx-test wait_states-itest

Tests are currently disabled in TSAN until
https://github.com/yugabyte/yugabyte-db/issues/20773 is fixed

Reviewers: arybochkin, esheng, asaha

Reviewed By: arybochkin

Subscribers: kannan, ybase, yql, bogdan

Differential Revision: https://phorge.dev.yugabyte.com/D31146

+++ src/yb/rpc/local_call.h
@@ -136,7 +136,7 @@ class LocalYBInboundCallTracker : public InboundCall::CallProcessedListener {
 
  private:
   std::mutex lock_;
-  std::unordered_map<uintptr_t, InboundCallWeakPtr> calls_being_handled_ GUARDED_BY(lock_);
+  std::unordered_map<int64_t, InboundCallWeakPtr> calls_being_handled_ GUARDED_BY(lock_);
 };
 
 } // namespace rpc
